refactor(sanity): extract common search token functions
