+++ packages/sanity/src/core/search/text-search/createTextSearch.ts
@@ -16,11 +16,9 @@ import {
   type TextSearchResponse,
   type TextSearchResults,
 } from '../common'
+import {prefixLast} from '../common/token'
 
 const DEFAULT_LIMIT = 1000
-const WILDCARD_TOKEN = '*'
-const NEGATION_TOKEN = '-'
-const TOKEN_REGEX = /(?:[^\s"]+|"[^"]*")+/g
 
 function normalizeSearchTerms(
   searchParams: string | SearchTerms,
@@ -84,40 +82,6 @@ export function getOrder(sort: SearchSort[] = []): TextSearchOrder[] {
     }),
     {},
   )
-}
-
-export function isNegationToken(token: string | undefined): boolean {
-  return typeof token !== 'undefined' && token.trim().at(0) === NEGATION_TOKEN
-}
-
-export function isPrefixToken(token: string | undefined): boolean {
-  return typeof token !== 'undefined' && token.trim().at(-1) === WILDCARD_TOKEN
-}
-
-export function isExactMatchToken(token: string | undefined): boolean {
-  return [token?.at(0), token?.at(-1)].every((character) => character === '"')
-}
-
-export function prefixLast(query: string): string {
-  const tokens = (query.match(TOKEN_REGEX) ?? []).map((token) => token.trim())
-
-  const finalIncrementalTokenIndex = tokens.findLastIndex(
-    (token) => !isNegationToken(token) && !isExactMatchToken(token),
-  )
-
-  const finalIncrementalToken = tokens[finalIncrementalTokenIndex]
-
-  if (tokens.length === 0) {
-    return WILDCARD_TOKEN
-  }
-
-  if (isPrefixToken(finalIncrementalToken) || typeof finalIncrementalToken === 'undefined') {
-    return tokens.join(' ')
-  }
-
-  const prefixedTokens = [...tokens]
-  prefixedTokens.splice(finalIncrementalTokenIndex, 1, `${finalIncrementalToken}${WILDCARD_TOKEN}`)
-  return prefixedTokens.join(' ')
 }
 
 export function getQueryString(
